% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{congress_hex}
\alias{congress_hex}
\title{Hexagon for US Congressional districts}
\format{A simple features object with 435 rows and 4 fields}
\source{
https://pitchinteractiveinc.github.io/tilegrams/
}
\usage{
congress_hex
}
\description{
A dataset containing geometries for plotting US Congressional districts as
hexagons, broken out by state.
}
\details{
Fields:

\itemize{
  \item `fips`: numeric. State FIPS code.
  \item `name`: character. State name in upper case.
  \item `state_abb`: charatcer. State abbreviation in upper case.
  \item `district`: numeric. District number. `1` for states with only one at-large
representative.
  \item `geometry`: POLYGON. Geometries for hexagons.
}
}
\keyword{datasets}
