% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize_points.R
\name{standard_hex_points}
\alias{standard_hex_points}
\title{Standard state hexagon points}
\usage{
standard_hex_points(hex_centroids_sf)
}
\description{
Given a simple feature of hexagon centroids, return
a simple features object with standardized x, y
}
