% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize_points.R
\name{standard_state_points}
\alias{standard_state_points}
\title{Standard state points}
\usage{
standard_state_points(real_centroids)
}
\arguments{
\item{real_centroids}{sf. An object containing the true Congressional
centroids.}
}
\description{
Given a simple feature of true centroids for a state, return a simple
features object with standardized x, y.
}
