% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{states_hex}
\alias{states_hex}
\title{Hexagon for all 50 US states}
\format{A simple features object with 50 rows and 3 fields}
\source{
https://pitchinteractiveinc.github.io/tilegrams/
}
\usage{
states_hex
}
\description{
A dataset containing geometries for plotting US states as hexagons.
}
\details{
Fields:

\itemize{
  \item `fips`: numeric. State FIPS code.
  \item `name`: character. State name in upper case.
  \item `state_abb`: charatcer. State abbreviation in upper case.
  \item `geometry`: POLYGON. Geometries for hexagons.
}
}
\keyword{datasets}
