% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_congress_hex.R
\name{match_districts_state}
\alias{match_districts_state}
\title{Match districts}
\usage{
match_districts_state(real_centroids, hex_centroids, iter = 10,
  wts = NULL, compare = FALSE)
}
\description{
Description
}
