% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize_points.R
\name{standard_hex_points}
\alias{standard_hex_points}
\title{Standard state hexagon points}
\usage{
standard_hex_points(hex_centroids)
}
\arguments{
\item{hex_centroids}{sf. An object containing the hexagon centroids.}
}
\description{
Given a simple feature of hexagon centroids for a state, return a simple
features object with standardized x, y.
}
